Feat: use '400 - Bad Request' repsonses; bunch of rearranging code

+++ app/admin/views.py
@@ -16,13 +16,37 @@ def admin_panel(username):
     form = AddCategoryForm()
     category_page = request.args.get('category_page', None, type=int)
     if not category_page:
-        category_pagination = Category.query.paginate(1, per_page=RESULT_COUNT, error_out=False)
+        category_pagination = (Category.query
+                               .paginate(1, per_page=RESULT_COUNT, error_out=False))
         categories = category_pagination.items
-        return render_template('admin/admin_panel.html', categories=categories, category_pagination=category_pagination, form=form, displays={"users_search_result_disp": "none", "book_categories_disp": "none", "add_category_disp": "none", "books_in_a_category_disp": "none"})
+        return render_template(
+            'admin/admin_panel.html',
+            categories=categories,
+            category_pagination=category_pagination,
+            form=form,
+            displays={
+                "users_search_result_disp": "none",
+                "book_categories_disp": "none",
+                "add_category_disp": "none",
+                "books_in_a_category_disp": "none"
+            }
+        )
     else:
-        category_pagination = Category.query.paginate(category_page, per_page=RESULT_COUNT, error_out=False)
+        category_pagination = (Category.query
+                               .paginate(category_page, per_page=RESULT_COUNT, error_out=False))
         categories = category_pagination.items
-        return render_template('admin/admin_panel.html', categories=categories, category_pagination=category_pagination, form=form, displays={"users_search_result_disp": "none", "book_categories_disp": "block", "add_category_disp": "none", "books_in_a_category_disp": "none"})
+        return render_template(
+            'admin/admin_panel.html',
+            categories=categories,
+            category_pagination=category_pagination,
+            form=form,
+            displays={
+                "users_search_result_disp": "none",
+                    "book_categories_disp": "block",
+                    "add_category_disp": "none",
+                    "books_in_a_category_disp": "none"
+            }
+        )
 
 
 @admin.route('/<username>/admin_panel/user_search', methods=['POST'])
@@ -36,11 +60,22 @@ def user_search(username):
             last_page = (len(users) - 1) // USERS_COUNT
             if len(users) % USERS_COUNT > 0:
                 last_page += 1
-            user_pagination = User.query.filter(User.username != current_user.username).paginate(1, per_page=USERS_COUNT, error_out=False)
+            user_pagination = (User.query
+                               .filter(User.username != current_user.username)
+                               .paginate(1, per_page=USERS_COUNT, error_out=False))
             if user_pagination.items:
                 users = user_pagination.items
                 pages_count = list(user_pagination.iter_pages())
-                return jsonify([dict(result=True, cur_page=1, pages_count=pages_count, page=last_page, id=user.id, username=user.username) for user in users])
+                return jsonify([
+                    dict(
+                        result=True,
+                        cur_page=1,
+                        pages_count=pages_count,
+                        page=last_page,
+                        id=user.id,
+                        username=user.username
+                    ) for user in users
+                ])
             else:
                 return jsonify([dict(result=False)])
         else:
@@ -49,7 +84,16 @@ def user_search(username):
             if user:
                 last_page = 1
                 pages_count = [1]
-                return jsonify([dict(result=True, cur_page=1, pages_count=pages_count, page=last_page, id=user.id, username=user.username)])
+                return jsonify([
+                    dict(
+                        result=True,
+                        cur_page=1,
+                        pages_count=pages_count,
+                        page=last_page,
+                        id=user.id,
+                        username=user.username
+                    )
+                ])
             else:
                 return jsonify([dict(result=False)])
     else:
@@ -60,20 +104,43 @@ def user_search(username):
 @admin_required
 @check_actual_password
 def get_user_search_page(page):
-    user_pagination = User.query.filter(User.username != current_user.username).paginate(page, per_page=USERS_COUNT, error_out=False)
+    user_pagination = (User.query
+                       .filter(User.username != current_user.username)
+                       .paginate(page, per_page=USERS_COUNT, error_out=False))
     pages_count = list(user_pagination.iter_pages())
     users = user_pagination.items
-    return jsonify([dict(cur_page=page, id=user.id, email=user.email, username=user.username, pages_count=pages_count, city=user.city, gender=user.gender, age=user.age, about_me=user.about_me) for user in users])
+    return jsonify([
+        dict(
+            cur_page=page,
+            id=user.id,
+            email=user.email,
+            username=user.username,
+            pages_count=pages_count,
+            city=user.city,
+            gender=user.gender,
+            age=user.age,
+            about_me=user.about_me
+        ) for user in users
+    ])
 
 
 @admin.route('/<username>/get_category_search_page/<int:page>', methods=['GET'])
 @admin_required
 @check_actual_password
 def get_category_search_page(username, page):
-    category_pagination = Category.query.paginate(page, per_page=RESULT_COUNT, error_out=False)
+    category_pagination = (Category.query
+                           .paginate(page, per_page=RESULT_COUNT, error_out=False))
     pages_count = list(category_pagination.iter_pages())
     categories = category_pagination.items
-    return jsonify([dict(username=current_user.username, cur_page=page, pages_count=pages_count, id=category.id, name=category.name) for category in categories])
+    return jsonify([
+        dict(
+            username=current_user.username,
+            cur_page=page,
+            pages_count=pages_count,
+            id=category.id,
+            name=category.name
+        ) for category in categories
+    ])
 
 
 @admin.route('/<username>/add-category', methods=['POST'])
@@ -82,32 +149,55 @@ def get_category_search_page(username, page):
 def add_category(username):
     form = AddCategoryForm()
     if form.validate_on_submit():
-        category = Category(name=str(form.category_name.data).strip().lower().replace("'", ""))
+        category = Category(
+                name=str(form.category_name.data).strip().lower().replace("'", "")
+        )
         database.session.add(category)
         database.session.commit()
         categories = Category.query.all()
         last_page = len(categories) // RESULT_COUNT
         if len(categories) % RESULT_COUNT > 0:
             last_page += 1
-        return redirect(url_for('.admin_panel', username=username, category_page=last_page))
-
+        return redirect(url_for(
+            '.admin_panel',
+            username=username,
+            category_page=last_page
+        ))
     category_page = request.args.get('category_page', 1, type=int)
-    category_pagination = Category.query.paginate(
-        category_page, per_page=RESULT_COUNT, error_out=False)
+    category_pagination = (Category.query
+                           .paginate(category_page, per_page=RESULT_COUNT, error_out=False))
     categories = category_pagination.items
-    return render_template('admin/admin_panel.html', categories=categories, form=form, category_pagination=category_pagination, displays={"users_search_result_disp": "none", "book_categories_disp": "block", "add_category_disp": "block", "books_in_a_category_disp": "none"})
+    return render_template(
+        'admin/admin_panel.html',
+        categories=categories,
+        form=form,
+        category_pagination=category_pagination,
+        displays={
+            "users_search_result_disp": "none",
+            "book_categories_disp": "block",
+            "add_category_disp": "block",
+            "books_in_a_category_disp": "none"
+        }
+    )
 
 
 @admin.route('/admin_panel/user_delete/<int:user_id>/<int:page>', methods=['GET'])
 @admin_required
 @check_actual_password
 def user_delete(user_id, page):
-    user = User.query.filter((User.id != current_user.id) & (User.id == user_id)).first()
+    user = User.query.filter(
+        (User.id != current_user.id) &
+        (User.id == user_id)
+    ).first()
+    if user is None:
+        return render_template("400.html")
     database.session.delete(user)
     database.session.commit()
     if User.query.filter(User.id != current_user.id).all():
         has_elems = True
-        user_pagination = User.query.filter(User.id != current_user.id).paginate(page, per_page=USERS_COUNT, error_out=False)
+        user_pagination = (User.query
+                           .filter(User.id != current_user.id)
+                           .paginate(page, per_page=USERS_COUNT, error_out=False))
         pages_count = list(user_pagination.iter_pages())
         if not user_pagination.items:
             page -= 1
@@ -115,7 +205,11 @@ def user_delete(user_id, page):
         page = 1
         pages_count = 1
         has_elems = False
-    return jsonify(dict(cur_page=page, pages_count=pages_count, has_elems=has_elems))
+    return jsonify(dict(
+        cur_page=page,
+        pages_count=pages_count,
+        has_elems=has_elems
+    ))
 
 
 @admin.route('/<username>/category_delete/<int:category_id>/<int:page>', methods=['GET'])
@@ -123,6 +217,8 @@ def user_delete(user_id, page):
 @check_actual_password
 def category_delete(username, category_id, page):
     category = Category.query.filter_by(id=category_id).first()
+    if category is None:
+        return render_template("400.html")
     database.session.delete(category)
     database.session.commit()
     if Category.query.all():
@@ -135,7 +231,12 @@ def category_delete(username, category_id, page):
         page = 1
         pages_count = 1
         has_elems = False
-    return jsonify(dict(username=current_user.username, cur_page=page, pages_count=pages_count, has_elems=has_elems))
+    return jsonify(dict(
+        username=current_user.username,
+        cur_page=page,
+        pages_count=pages_count,
+        has_elems=has_elems
+    ))
 
 
 @admin.route('/<username>/search_books_on_admin_panel/<int:category_id>', methods=['GET', 'POST'])
@@ -143,15 +244,22 @@ def category_delete(username, category_id, page):
 @check_actual_password
 def search_books_on_admin_panel(username, category_id):
     category = Category.query.filter_by(id=category_id).first()
+    if category is None:
+        return render_template("400.html")
     if request.method == "POST":
         result = str(request.form.get('search_result')).strip().lower()
         if result == '*':
             books = category.books.all()
         else:
-            books = category.books.filter((Book.name.like("%{}%".format(result))) | (
-                Book.author.like("%{}%".format(result)))).all()
+            books = category.books.filter(
+                (Book.name.like("%{}%".format(result))) |
+                (Book.author.like("%{}%".format(result)))
+            ).all()
         search_result_fin = []
-        [search_result_fin.append(value) for value in books if value and value not in search_result_fin]
+        # [search_result_fin.append(value) for value in books if value and value not in search_result_fin]
+        for value in books:
+            if value and not value in search_result_fin:
+                search_result_fin.append(value)
         books = search_result_fin
         if books:
             pages = len(books) // RESULT_COUNT
@@ -170,7 +278,10 @@ def search_books_on_admin_panel(username, category_id):
                 pages_count = list(range(1, pages + 1))      
                              
             # удаление прошлых результатов поиска из бд
-            database.session.query(SearchResult).filter_by(searcher_id=current_user.id).delete()
+            (database.session
+                    .query(SearchResult)
+                    .filter_by(searcher_id=current_user.id)
+                    .delete())
                  
             books_grades = list()
             for book in books:
@@ -184,16 +295,46 @@ def search_books_on_admin_panel(username, category_id):
             database.session.commit()
             if len(books) > RESULT_COUNT:
                 books = books[:RESULT_COUNT]
-            return jsonify([dict(has_books=True, pages=pages, cur_page=1, pages_count=pages_count, username=current_user.username, category=category.id, id=book.id, name=book.name, author=book.author, release_date=book.release_date, grade=book_grade) for book, book_grade in zip(books, books_grades)])
+            return jsonify([
+                dict(
+                    has_books=True,
+                    pages=pages,
+                    cur_page=1,
+                    pages_count=pages_count,
+                    username=current_user.username,
+                    category=category.id,
+                    id=book.id,
+                    name=book.name,
+                    author=book.author,
+                    release_date=book.release_date,
+                    grade=book_grade
+                ) for book, book_grade in zip(books, books_grades)
+            ])
         else:
             return jsonify([dict(has_books=False)])
 
     elif page := request.args.get('page', None, type=int):
-        cur_result_pagination = SearchResult.query.filter_by(searcher_id=current_user.id).order_by(SearchResult.id).paginate(page, per_page=RESULT_COUNT, error_out=False)
+        cur_result_pagination = (SearchResult.query
+                                 .filter_by(searcher_id=current_user.id)
+                                 .order_by(SearchResult.id)
+                                 .paginate(page, per_page=RESULT_COUNT, error_out=False))
         pages_count = list(cur_result_pagination.iter_pages())
         cur_result = cur_result_pagination.items
         if cur_result:
-            return jsonify([dict(has_books=True, cur_page=page, pages_count=pages_count, username=current_user.username, category=category.id, id=book.id, name=book.name, author=book.author, release_date=book.release_date, grade=book.grade) for book in cur_result])
+            return jsonify([
+                dict(
+                    has_books=True,
+                    cur_page=page,
+                    pages_count=pages_count,
+                    username=current_user.username,
+                    category=category.id,
+                    id=book.id,
+                    name=book.name,
+                    author=book.author,
+                    release_date=book.release_date,
+                    grade=book.grade
+                ) for book in cur_result
+            ])
         else:
             return jsonify([dict(has_books=False)])
     return render_template('500.html')
@@ -204,6 +345,8 @@ def search_books_on_admin_panel(username, category_id):
 @check_actual_password
 def del_book(username, category_id, book_id, page):
     book = Book.query.filter_by(id=book_id).first()
+    if book is None:
+        return render_template("400.html")
     database.session.delete(book)
     book_for_search_result = SearchResult.query.filter_by(searcher_id=current_user.id).filter_by(id=book_id).first()
     database.session.delete(book_for_search_result)
@@ -218,7 +361,13 @@ def del_book(username, category_id, book_id, page):
         page = 1
         pages_count = 1
         has_elems = False
-    return jsonify(dict(cur_page=page, pages_count=pages_count, has_elems=has_elems, username=current_user.username, category=category_id))
+    return jsonify(dict(
+        cur_page=page,
+        pages_count=pages_count,
+        has_elems=has_elems,
+        username=current_user.username,
+        category=category_id
+    ))
 
 
 @admin.route('/<username>/change_book_info/<int:book_id>', methods=['GET', 'POST'])
@@ -227,19 +376,35 @@ def del_book(username, category_id, book_id, page):
 def change_book_info(username, book_id):
     if current_user.username != username:
         return render_template('403.html')
-    if book := Book.query.filter_by(id=book_id).first():
-        pagination = Category.query.paginate(1, per_page=CATEGORIES_COUNT, error_out=False)
-        categories = pagination.items
-        form = ChangeBookInfoForm(book=book)
-        if form.validate_on_submit():
-            category = Category.query.filter_by(name=request.form.get('category')).first()
-            if cover := bytes(request.files['cover'].read()):
-                book.cover=cover
-            book.isbn=form.isbn.data.strip(); book.name=form.name.data.strip().lower().replace("'", ""); book.author=form.author.data.strip().lower(); book.publishing_house=form.publishing_house.data.strip()
-            book.description=request.form.get('description').strip(); book.release_date=form.release_date.data; book.count_of_chapters=form.chapters_count.data
-            book.category=category; book.user=current_user._get_current_object()
-            database.session.add(book)
-            database.session.commit()
-            return redirect(url_for('main.book_page', book_id=book.id))
-        return render_template('admin/change_book_info.html', form=form, categories=categories, pagination=pagination, book=book, range=range, len=len)
-    return render_template('404.html')
+    book = Book.query.filter_by(id=book_id).first()
+    if book is None:
+        return render_template("400.html")
+    pagination = Category.query.paginate(1, per_page=CATEGORIES_COUNT, error_out=False)
+    categories = pagination.items
+    form = ChangeBookInfoForm(book=book)
+    if form.validate_on_submit():
+        category = Category.query.filter_by(name=request.form.get('category')).first()
+        if cover := bytes(request.files['cover'].read()):
+            book.cover = cover
+        book.isbn = form.isbn.data.strip()
+        book.name = form.name.data.strip().lower().replace("'", "")
+        book.author = form.author.data.strip().lower()
+        book.publishing_house = form.publishing_house.data.strip()
+        book.description = request.form.get('description').strip()
+        book.release_date = form.release_date.data
+        book.count_of_chapters = form.chapters_count.data
+        book.category = category
+        book.user = current_user._get_current_object()
+        database.session.add(book)
+        database.session.commit()
+        return redirect(url_for('main.book_page', book_id=book.id))
+    return render_template(
+        'admin/change_book_info.html',
+        form=form,
+        categories=categories,
+        pagination=pagination,
+        book=book,
+        range=range,
+        len=len
+    )
+
